+++ lib/generate/json2dart/json2dart_command.dart
@@ -225,11 +225,11 @@ class Json2DartCommand extends Command {
       }
     }
 
-    print('Execute morpheme format................');
+    printMessage('Execute morpheme format................');
     if (format.isNotEmpty) await ModularHelper.format(format);
-    print('Execute morpheme fix................');
+    printMessage('Execute morpheme fix................');
     if (fix.isNotEmpty) await ModularHelper.fix(fix);
-    print('Execute morpheme format................');
+    printMessage('Execute morpheme format................');
     if (format.isNotEmpty) await ModularHelper.format(format);
 
     StatusHelper.success('morpheme json2dart');